+++ components/SuggestedNodeComponent.tsx
@@ -13,8 +13,8 @@ const handleStyle: React.CSSProperties = {
 
 type PineNodeProps = NodeProps<SuggestedNodeData>;
 
-const onSuggestedNodeClick = (session: Session, pine: string) => {
-  session.pipeAndUpdateExpression(pine, true);
+const onSuggestedNodeClick = async (session: Session, pine: string) => {
+  await session.pipeAndUpdateExpression(pine, true);
 };
 
 const SuggestedNodeComponent: React.FC<PineNodeProps> = ({ data }) => {